feat: move changes to main classes
