+++ sql/src/main/kotlin/com/smeup/dbnative/sql/Native2SQLAdapter.kt
@@ -4,31 +4,30 @@ import com.smeup.dbnative.file.Record
 import com.smeup.dbnative.model.FileMetadata
 import com.smeup.dbnative.model.Field
 import java.lang.Exception
-import java.sql.Connection
 
 enum class PositioningMethod {
     SETLL, SETGT;
 }
 
-enum class ReadMethod(val forward: Boolean){
+enum class ReadMethod(val forward: Boolean) {
     READE(true), READPE(false), READP(false), READ(true), CHAIN(true);
 }
 
-enum class SortOrder(val symbol: String){
+enum class SortOrder(val symbol: String) {
     ASCEDING("ASC"), DESCENDING("DESC");
 }
 
-class PositioningInstruction(val method: PositioningMethod, val keys: List<String>){
+class PositioningInstruction(val method: PositioningMethod, val keys: List<String>) {
     init {
-        require(keys.isNotEmpty()){
+        require(keys.isNotEmpty()) {
             "No keys specified for positioning instruction $method"
         }
     }
 }
 
-class ReadInstruction(var method: ReadMethod, var keys: List<String>){
-    constructor(method: ReadMethod): this(method, emptyList()){
-        require(admitEmptyKeys()){
+class ReadInstruction(var method: ReadMethod, var keys: List<String>) {
+    constructor(method: ReadMethod) : this(method, emptyList()) {
+        require(admitEmptyKeys()) {
             "Keys are mandatory for read instruction $method "
         }
     }
@@ -43,65 +42,57 @@ class Native2SQL(val fileMetadata: FileMetadata) {
     /**
      * Build values replacements for store procedures settings empty numerics values as 0
      */
-    private fun buildRepalcements(values: List<String>): MutableList<String> {
-        val result = mutableListOf<String>()
-        values.forEachIndexed { index, s ->
-            if (s.isNullOrEmpty()) {
+    private fun buildReplacements(values: List<String>): MutableList<String> {
+        return values.mapIndexed { index, s ->
+            s.ifEmpty {
                 if (isNumeric(fileMetadata.fileKeys[index])) {
-                    result.add("0")
+                    "0"
                 } else {
-                    result.add("");
+                    ""
                 }
-            } else {
-                result.add(s)
             }
-        }
-        return result
+        }.toMutableList()
     }
 
-    private fun isNumeric(fieldName: String ):Boolean {
+    private fun isNumeric(fieldName: String): Boolean {
         val field = fileMetadata.fields.find { it.name == fieldName }
-        if (field != null) {
-            return field.numeric
-        } else {
-            return false
-        }
+        return field?.numeric ?: false
     }
 
-    private fun checkPositioning(){
-        requireNotNull(lastPositioningInstruction){
+    private fun checkPositioning() {
+        requireNotNull(lastPositioningInstruction) {
             "No positioning instruction found"
         }
     }
 
-    private fun checkKeys(keys: List<String>){
-        require(fileMetadata.fileKeys.size > 0){
+    private fun checkKeys(keys: List<String>) {
+        require(fileMetadata.fileKeys.size > 0) {
             "No keys specified in metadata"
         }
-        require(keys.size <= fileMetadata.fileKeys.size){
+        require(keys.size <= fileMetadata.fileKeys.size) {
             "Number of metadata keys $fileMetadata.fileKeys less than number of positioning/read keys $keys"
         }
     }
 
-    private fun checkRead(){
-        requireNotNull(lastReadInstruction){
+    private fun checkRead() {
+        requireNotNull(lastReadInstruction) {
             "No read instruction found"
         }
     }
 
-    private fun checkReadKeys(){
+    private fun checkReadKeys() {
         checkRead()
-        require(!lastReadInstruction!!.keys.isNullOrEmpty()){
+        require(lastReadInstruction!!.keys.isNotEmpty()) {
             "No keys specified for read instruction ${lastReadInstruction!!.method}"
         }
     }
 
-    private fun checkInstructions(){
+    private fun checkInstructions() {
         //checkPositioning()
         checkRead()
     }
 
-    fun setPositioning(method: PositioningMethod, keys: List<String>){
+    fun setPositioning(method: PositioningMethod, keys: List<String>) {
         checkKeys(keys)
         lastPositioningInstruction = PositioningInstruction(method, keys)
         lastReadInstruction = null
@@ -112,18 +103,18 @@ class Native2SQL(val fileMetadata: FileMetadata) {
      * - If passed key are less then file metadata keys fill missing var with empty values
      * - If a key is numeric and is empty, set its value to 0
      */
-    private fun checkEmptyKeys(keys: List<String>): List<String>  {
+    private fun checkEmptyKeys(keys: List<String>): List<String> {
         val result = mutableListOf<String>()
 
         result.addAll(keys)
         val lostKeys = fileMetadata.fileKeys.size - result.size
         repeat(lostKeys) {
-            result.add("");
+            result.add("")
         }
 
         val checkedResult = mutableListOf<String>()
         result.forEachIndexed { index, s ->
-            if (s.isNullOrEmpty() ) {
+            if (s.isNullOrEmpty()) {
                 if (isNumeric(fileMetadata.fileKeys.get(index))) {
                     checkedResult.add("0")
                 } else {
@@ -140,17 +131,18 @@ class Native2SQL(val fileMetadata: FileMetadata) {
     /*
      * @return true if read method need new query execution
      */
-    fun setRead(method: ReadMethod, keys: List<String>? = null): Boolean{
+    fun setRead(method: ReadMethod, keys: List<String>? = null): Boolean {
         checkKeys(keys ?: emptyList())
         var executeQuery = false
-        when(method){
-            ReadMethod.READPE, ReadMethod.READE ->{
-                val coherent = keys?.let{isCoherent(keys)} ?: true
+        when (method) {
+            ReadMethod.READPE, ReadMethod.READE -> {
+                val coherent = keys?.let { isCoherent(keys) } ?: true
                 //Test to remove on fully supported operations
-                require(coherent){
+                require(coherent) {
                     "Uncoherent read not yet managed"
                 }
             }
+
             ReadMethod.READP -> checkPositioning()
             ReadMethod.CHAIN -> {
                 lastPositioningInstruction = null
@@ -160,31 +152,32 @@ class Native2SQL(val fileMetadata: FileMetadata) {
         require(lastReadInstruction == null || method == ReadMethod.CHAIN || method == lastReadInstruction!!.method) {
             "read operation " + method + " is allowed immediatly after positioning or after a same method read instruction"
         }
-        if(lastReadInstruction == null){
+        if (lastReadInstruction == null) {
             executeQuery = true
         }
-        lastReadInstruction = ReadInstruction(method, keys ?:emptyList())
+        lastReadInstruction = ReadInstruction(method, keys ?: emptyList())
         return executeQuery
     }
 
-    fun clear(){
+    fun clear() {
         lastReadInstruction = null
         lastPositioningInstruction = null
     }
 
-    fun getLastKeys() = lastReadInstruction?.keys ?:lastPositioningInstruction?.keys ?: throw Exception("Keys not yet set")
+    fun getLastKeys() =
+        lastReadInstruction?.keys ?: lastPositioningInstruction?.keys ?: throw Exception("Keys not yet set")
 
-    fun isLastOperationSet()=lastReadInstruction == null
+    fun isLastOperationSet() = lastReadInstruction == null
 
     fun lastReadMatchRecord(record: Record): Boolean {
 
-        if(lastReadInstruction!!.keys.isEmpty()){
+        if (lastReadInstruction!!.keys.isEmpty()) {
             return true
         }
 
         lastReadInstruction!!.keys.mapIndexed { index, value ->
             val keyname = fileMetadata.fileKeys.get(index)
-            if(record[keyname]?.trim() != value.trim()){
+            if (record[keyname]?.trim() != value.trim()) {
                 return false
             }
         }
@@ -192,12 +185,12 @@ class Native2SQL(val fileMetadata: FileMetadata) {
     }
 
 
-    fun isCoherent(newKeys: List<String>): Boolean{
+    fun isCoherent(newKeys: List<String>): Boolean {
         //checkPositioning()
         if (lastPositioningInstruction != null) {
             return newKeys.isEmpty() ||
                     if (newKeys.size <= lastPositioningInstruction!!.keys.size &&
-                        newKeys.size <= lastReadInstruction?.keys?.size ?: newKeys.size
+                        newKeys.size <= (lastReadInstruction?.keys?.size ?: newKeys.size)
                     ) {
                         newKeys.forEachIndexed() { index, value ->
                             if (lastPositioningInstruction!!.keys.get(index) != value) {
@@ -211,57 +204,71 @@ class Native2SQL(val fileMetadata: FileMetadata) {
 
     private fun getSortOrder(): SortOrder {
         checkInstructions()
-        return if(lastReadInstruction!!.method.forward) SortOrder.ASCEDING else SortOrder.DESCENDING
+        return if (lastReadInstruction!!.method.forward) SortOrder.ASCEDING else SortOrder.DESCENDING
     }
 
-    private fun getComparison(): Pair<Comparison, Comparison>{
+    private fun getComparison(): Pair<Comparison, Comparison> {
         checkInstructions()
         return if (lastPositioningInstruction == null) {
             return Pair(Comparison.EQ, Comparison.GT)
-        } else if(lastReadInstruction!!.method.forward){
-            when(lastPositioningInstruction!!.method){
-                        PositioningMethod.SETLL -> return Pair(Comparison.EQ, Comparison.GT)
-                        PositioningMethod.SETGT -> return Pair(Comparison.GT, Comparison.GT)
+        } else if (lastReadInstruction!!.method.forward) {
+            when (lastPositioningInstruction!!.method) {
+                PositioningMethod.SETLL -> return Pair(Comparison.EQ, Comparison.GT)
+                PositioningMethod.SETGT -> return Pair(Comparison.GT, Comparison.GT)
             }
-        }
-        else{
-            when(lastPositioningInstruction!!.method){
+        } else {
+            when (lastPositioningInstruction!!.method) {
                 PositioningMethod.SETLL -> return Pair(Comparison.LT, Comparison.LT)
                 PositioningMethod.SETGT -> return Pair(Comparison.LE, Comparison.LT)
             }
         }
     }
 
-    private fun getSQLOrderByClause(): String{
+    private fun getSQLOrderByClause(): String {
         val sortOrder = getSortOrder()
-        var result = "ORDER BY"
-
-        fileMetadata.fileKeys.forEach() {
-            result += " \"" + it + "\" " + sortOrder.symbol +","
-        }
-        return result.removeSuffix(",")
+        return fileMetadata.fileKeys.joinToString(
+            prefix = "ORDER BY ",
+            separator = ", "
+        ) { "\"$it\" ${sortOrder.symbol}" }
     }
 
-    fun getReadSqlStatement(): Pair<String, List<String>>{
+    fun getReadSqlStatement(): Pair<String, List<String>> {
         checkPositioning()
-        return Pair(getSQL(fileMetadata.fields, fileMetadata.fileKeys.subList(0, lastPositioningInstruction!!.keys.size), Comparison.EQ, fileMetadata.tableName), lastPositioningInstruction!!.keys)
+        return Pair(
+            getSQL(
+                fileMetadata.fields,
+                fileMetadata.fileKeys.subList(0, lastPositioningInstruction!!.keys.size),
+                Comparison.EQ,
+                fileMetadata.tableName
+            ), lastPositioningInstruction!!.keys
+        )
     }
 
-    fun getSQLSatement(): Pair<String, List<String>>{
-        when(lastReadInstruction!!.method){
-            ReadMethod.CHAIN ->{
+    fun getSQLStatement(): Pair<String, List<String>> {
+        when (lastReadInstruction!!.method) {
+            ReadMethod.CHAIN -> {
                 checkReadKeys()
-                return Pair(getSQL(fileMetadata.fields, fileMetadata.fileKeys.subList(0, lastReadInstruction!!.keys.size), Comparison.EQ, fileMetadata.tableName), lastReadInstruction!!.keys)
+                return Pair(
+                    getSQL(
+                        fileMetadata.fields,
+                        fileMetadata.fileKeys.subList(0, lastReadInstruction!!.keys.size),
+                        Comparison.EQ,
+                        fileMetadata.tableName
+                    ), lastReadInstruction!!.keys
+                )
             }
+
             ReadMethod.READ -> {
                 checkRead()
                 return getReadCoherentSql(true)
             }
+
             ReadMethod.READP -> {
                 checkPositioning()
                 checkRead()
                 return getCoherentSql(true)
             }
+
             else -> {
                 //checkPositioning()
                 checkReadKeys()
@@ -283,53 +290,29 @@ class Native2SQL(val fileMetadata: FileMetadata) {
     FROM EMPL0F
     WHERE WORKDEPT >= 'C01'
      */
-    private fun getReadCoherentSql(fullUnion: Boolean = false): Pair<String, MutableList<String>> {
-        var queries = mutableListOf<String>()
-        var replacements = mutableListOf<String>()
-        if (lastPositioningInstruction == null) {
-            var columns = ""
-            fileMetadata.fields.forEachIndexed { index, k ->
-                run {
-                    columns += "\"" + k.name + "\", "
-                }
-            }
-            return Pair("SELECT " + columns.removeSuffix(", ") + "FROM \"${fileMetadata.tableName}\"", replacements)
-        } else {
-            var columns = ""
-            fileMetadata.fields.forEachIndexed { index, k ->
-                run {
-                    columns += "\"" + k.name + "\", "
-                }
+    private fun getReadCoherentSql(fullUnion: Boolean = false): Pair<String, List<String>> {
+        val columns = fileMetadata.fields.joinToString(", ") { "\"${it.name}\"" }
+        val tableName = "\"${fileMetadata.tableName}\""
+        val queries = mutableListOf<String>()
+        val replacements = mutableListOf<String>()
+
+        lastPositioningInstruction?.let {
+            val conditions = it.keys.joinToString(" AND ") { key ->
+                "\"${fileMetadata.fileKeys[it.keys.indexOf(key)]}\" >= ?"
             }
 
-            var value = ""
-            val keys = lastPositioningInstruction?.keys
-            keys?.forEachIndexed { index, k ->
-                run {
-                    value += "\"" + fileMetadata.fileKeys[index] + "\" >= ? AND "
-                }
-            }
-
-            queries.add(
-                "SELECT " + columns.removeSuffix(", ") + " FROM \"${fileMetadata.tableName}\" WHERE " + value.removeSuffix(
-                    " AND "
-                )
-            )
-
-            replacements.addAll(buildRepalcements(lastPositioningInstruction!!.keys))
-            /*
-            val lostKeys = fileMetadata.fileKeys.size - lastPositioningInstruction!!.keys.size
-            repeat(lostKeys) {
-                replacements.add("");
-            }
-            */
-            return Pair(queries.joinToString() + " " + getSQLOrderByClause(), replacements)
+            queries.add("SELECT $columns FROM $tableName WHERE $conditions")
+            replacements.addAll(buildReplacements(it.keys))
+        } ?: run {
+            return Pair("SELECT $columns FROM $tableName", replacements)
         }
+
+        return Pair(queries.joinToString(" ") + " " + getSQLOrderByClause(), replacements)
     }
 
-    private fun getCoherentSql(fullUnion: Boolean = false):Pair<String, List<String>>{
+    private fun getCoherentSql(fullUnion: Boolean = false): Pair<String, List<String>> {
         val queries = mutableListOf<String>()
-        var replacements = mutableListOf<String>()
+        val replacements = mutableListOf<String>()
         val comparison = getComparison()
 
         if (lastPositioningInstruction == null) {
@@ -345,9 +328,13 @@ class Native2SQL(val fileMetadata: FileMetadata) {
                     value += "\"" + k + "\" " + Comparison.EQ.symbol + " ? AND "
                 }
             }
-            replacements.addAll(buildRepalcements(lastReadInstruction!!.keys))
+            replacements.addAll(buildReplacements(lastReadInstruction!!.keys))
 
-            return Pair("SELECT " + columns.removeSuffix(", ") + "FROM \"${fileMetadata.tableName}\" WHERE " + value.removeSuffix(" AND "), replacements)
+            return Pair(
+                "SELECT " + columns.removeSuffix(", ") + "FROM \"${fileMetadata.tableName}\" WHERE " + value.removeSuffix(
+                    " AND "
+                ), replacements
+            )
         } else {
             if (lastPositioningInstruction!!.keys.size == 1) {
                 queries.add(
@@ -358,7 +345,7 @@ class Native2SQL(val fileMetadata: FileMetadata) {
                         fileMetadata.tableName
                     )
                 )
-                replacements.addAll(buildRepalcements(lastPositioningInstruction!!.keys))
+                replacements.addAll(buildReplacements(lastPositioningInstruction!!.keys))
             } else {
                 val limit = if (fullUnion) 1 else 2
                 for (i in lastPositioningInstruction!!.keys.size downTo limit) {
@@ -370,7 +357,7 @@ class Native2SQL(val fileMetadata: FileMetadata) {
                             fileMetadata.tableName
                         )
                     )
-                    replacements.addAll(buildRepalcements(lastPositioningInstruction!!.keys.subList(0, i)))
+                    replacements.addAll(buildReplacements(lastPositioningInstruction!!.keys.subList(0, i)))
                 }
             }
         }
@@ -378,35 +365,28 @@ class Native2SQL(val fileMetadata: FileMetadata) {
     }
 }
 
-private fun getSQL(fields: List<Field>, keys: List<String>, comparison: Comparison, tableName: String): String{
-    var columns = ""
-    fields.forEachIndexed { index, k ->
-        run {
-            columns += "\"" + k.name + "\", "
-        }
-    }
+private fun getSQL(fields: List<Field>, keys: List<String>, comparison: Comparison, tableName: String): String {
 
-    var value = ""
-    keys.forEachIndexed { index, k ->
-        run {
-            value += "\"" + k + "\" " + (if (index < keys.size - 1) Comparison.EQ.symbol else  comparison.symbol) + " ? AND "
-        }
-    }
-    return "(SELECT " + columns.removeSuffix(", ")+ " FROM \"$tableName\" WHERE " + value.removeSuffix("AND ") + ")"
+    val columns = fields.joinToString(", ") { "\"${it.name}\"" }
+
+    val conditions = keys.mapIndexed { index, key ->
+        "\"$key\" ${if (index < keys.size - 1) Comparison.EQ.symbol else comparison.symbol} ?"
+    }.joinToString(" AND ")
+
+    return "SELECT $columns FROM \"$tableName\" WHERE $conditions"
 }
 
 
-
-fun main(){
+fun main() {
     var fields = listOf(Field("Regione", ""), Field("Provincia", ""), Field("Comune", ""))
     var fieldsKeys = listOf("Regione", "Provincia", "Comune")
     var metadata = FileMetadata("test", "rld_comuni", fields, fieldsKeys)
-    val adapter = Native2SQL(metadata).apply{
+    val adapter = Native2SQL(metadata).apply {
         setPositioning(PositioningMethod.SETLL, listOf("Lombardia", "Brescia", "Erbusco"))
         println(isCoherent(listOf("Lombardia", "Brescia", "Erbusco")))
         setRead(ReadMethod.READE, listOf("Lombardia", "Brescia", "Erbusco"))
     }
-    adapter.getSQLSatement().let{
+    adapter.getSQLStatement().let {
         println(it.first)
         println(it.second)
     }
